busybox: add suspend-actions script for the ccimx8x and ccimx6ul
This script behaves the same way as our custom standby script, with a few
modifications, for example:

 * instead of using initscripts to stop/start daemons, it uses systemctl calls
 * save variables that are needed when resuming in /tmp, because the script is
executed twice instead of once (when suspending and when resuming)
 * remove the locking/unlocking of the critical section and the actual state
change, since that is handled by systemd-suspend now

Also, add a different standby script for when systemd is enabled. Said script
acts as a wrapper for systemd's suspend mechanism.

https://jira.digi.com/browse/DEL-6415

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>