imx-atf: remove our .bb file and add a .bbappend
Bitbake was using the imx-atf recipe inside of meta-freescale instead of ours.
Said recipe was compiling a different tag, causing our kernel to stop loading
when using our generated imx-boot images.

Remove our .bb file and add a .bbappend that changes the source code revision
to the tag compatible with the rest of the imx-boot files.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>