networkmanager: disable terminal colors by default for nmcli
Since the version 1.12, NetworkManager enables the colors in the
terminal output by default. This pretty output is not compatible
with several terminal emulators and makes it difficult to parse
the output of the CLI commands.
This patch uses the standard linux util called 'terminal-colors.d'
to disable the colors by default in the NetworkManager CLI interpeter.

https://jira.digi.com/browse/DEL-6473

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>