greengrass: add systemd support
Since our initscript contains logic to dynamically generate the path containing
the greengrass daemon (it changes depending on the version), keep it in the
rootfs. Create a symlink to it inside of /etc/init.d for sysvinit images and
create a systemd service wrapper that calls it on systemd images.

https://jira.digi.com/browse/DEL-6510

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>