imx-usb-loader: use the version that was used in older versions of DEY
In the update to thud, the recipe for this package was updated in
meta-freescale with a newer revision that, among other things, breaks
compatibility with the parameters we use to recover closed devices via USB.
Even if we change these parameters to obtain a configuration equivalent to the
one we've always had, closed devices still won't boot because the internal
logic of the tool has changed a lot.

Instead of patching the tool for this very specific use-case, revert to the
version we've been using in DEY 2.2 and 2.4.

https://jira.digi.com/browse/DEL-6520

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>