python-numpy: redefine the target-specific RDEPENDS items
For some reason, even though the general RDEPENDS for this package is correct,
all items are overwritten by the empty class-native version of RDEPENDS, even
when building for the target. Because of this, whenever this python module is
installed and used, it won't work.

Redefine the dependencies needed by the target until the issue is fixed
upstream.

https://jira.digi.com/browse/DEL-6553

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>