xserver-xorg: add poky SIGIO patch that was removed from upstream thud branch
This patch is included in the rocko and sumo branches of poky to avoid errors
when starting an xserver using DRI, but it was removed in the thud branch in
favor of newer input methods other than SIGIO. SIGIO is still supported, but
not explicitly in the xserver configuration.

Since the vivante driver on the i.MX6 is using DRI and, by extension, SIGIO
(even though there is no way to configure xserver to enable SIGIO support), the
only way to make xserver start correctly is by removing the SIGIO support check
in the code.

https://jira.digi.com/browse/DEL-6565

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>