imx-atf: update to v2.0
This is the version used in meta-fsl-bsp-release branch sumo-4.14.98-2.0.0_ga.
When using v1.5, we used a .bbappend to change the revision of the v1.5 recipe
in meta-freescale. Since we're upgrading to a version that isn't supported in
the other layers yet, remove the .bbappend and add the entire v2.0 recipe.

https://jira.digi.com/browse/DEL-6603

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>