u-boot: fix the use of boot image recipe and prefix
A variable called IMAGE_BOOTLOADER was being used without distinction for
referring to two different things:
 - the recipe that builds the bootable artifacts
 - the prefix of those artifacts

The value of this is "u-boot" for most platforms, but "imx-boot" for the
CC8X based platforms.

The name of the variable is misleading, so this commit splits it into two:
 - BOOTLOADER_IMAGE_RECIPE, to refer to the recipe
 - UBOOT_PREFIX, to refer to the prefix of the bootable artifact

With the separation, the variable UBOOT_SYMLINK becomes a generic formed
one, so it is moved to digi-defaults.inc.

While on it, fix the image_types_digi.bbclass which was not making use of
the original variable to establish all the dependencies.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>