qualcomm.sh: remount as writteable if '/' is ro
On some scenarios, like when booting from NFS, the root filesystem is not
writteable. When this happens the qualcomm.sh is not able to write the MAC
file which so the wireless driver fails to load.

The arguments that U-Boot uses when booting from NFS using dboot result in
the root filesystem being mounted as read-only, but at some point it is
remounted as read/write. This happens after trying to load the wireless
driver, though.

Currently this is the only scenario in which we can reproduce the problem,
and there are several other processes in DEY that require a writeable root
filesystem (wireless SSID editing, Dropbear SSH key-gen and u-boot-fw-utils)
so remount the rootfs as writteable as a workaround.

While on it improve the script a bit to notify on error trying to create
this file, just in case this happens in other scenarios.

https://jira.digi.com/browse/DEL-5854

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>