udev-extraconf: update mount scripts to support systemd
The current scripts used the standard mount binary at /bin/mount
for mounting the partitions. Systemd however seems to have a monitor
that eventually umounts such partitions if not mounted by systemd's
specific command 'systemd-mount`.

For platforms using systemd, we need to use this binary together with
parameter --no-block so that partitions are mounted correctly during
the boot process.

This commit merges the two scripts for Digi-handled partitions into
just one:

	mount_bootparts.sh
                          \
                           +--> mount_digiparts.sh
                          /
        mount_partition.sh

The merged script:
 - checks if running with systemd, to use its binary.
 - checks if running with busybox, to not use the unsupported '-o'
   attribute.
 - checks if mounting the 'linux' partition, to mount it read-only.
 - checks for already-attached UBI devices, to avoid re-attaching.

This commit also combines the Digi-handled partitions (linux, update)
into the same udev rule, to simplify it, and breaks the rules lines
for readability.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6744