udev-extraconf: mount.sh with skip of already-mounted partitions
Default poky's version of the 'mount.sh' script doesn't account
for the possibility of udev calling twice the script for the 'add'
action. This has been observed on USB sticks connected to a USB hub
and, although the script eventually fails, for already mounted
partitions, in that case the script tries to remove the created
mountpoint, when it shouldn't.

This patch checks for already mounted partitions earlier (before
creating the mount point) and returns with no actions.

The patch that it adds (with respect to poky's thud default) is
the following:

  diff --git a/meta/recipes-core/udev/udev-extraconf/mount.sh b/meta/recipes-core/udev/udev-extraconf/mount.sh
  index 3ee67b13183a..e6882a69c223 100644
  --- a/meta/recipes-core/udev/udev-extraconf/mount.sh
  +++ b/meta/recipes-core/udev/udev-extraconf/mount.sh
  @@ -38,6 +38,12 @@ done
   automount_systemd() {
       name="`basename "$DEVNAME"`"

  +    # Skip already mounted partitions
  +    if [ -f /run/systemd/transient/run-media-$name.mount ]; then
  +        logger "mount.sh/automount" "/run/media/$name already mounted"
  +        return
  +    fi
  +
       # Skip the partition which are already in /etc/fstab
       grep "^[[:space:]]*$DEVNAME" /etc/fstab && return
       for n in LABEL PARTLABEL UUID PARTUUID; do

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6759