qualcomm.sh: generalize to select correct board data file
Generalize script to select the correct board data file to load into the
wireless chip depending on the Regulatory Domain configuration whether it
is a variation of 'bdwlan30.bin' or 'fakeboar.bin'.

If 'bdwlan30_US.bin' or 'bdwlan30_World.bin' files are found in the firmware
directory, the script will create two symlinks, 'bdwlan30.bin and 'utfbd30.bin',
pointing to either one of those files.

If no 'bdwlan30_*.bin' files are found, but 'fakeboar_US.bin' or
'fakeboar_World.bin' is, it will create a single symlink, 'fakeboar.bin',
pointing to either one of those files. In the case a QCA6574 community driver
is loaded, 'board.bin' symlink will be updated.

If no 'fakeboard_*.bin' files are found either, the script will exit.

https://jira.digi.com/browse/DEL-6773

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>