qtbase: apply imxgpu2d patches to imx8mn platforms as well
The .bbappend file in meta-freescale only applies these patches to platforms
with imxqpu2d, but they are necessary to build framebuffer images for the
ccimx8mn. Using the same approach as NXP in meta-fsl-bsp-release (applying the
patches to imxgpu platforms) duplicates the patches for platforms that are both
imxgpu and imxgpu2d, so append them for imx8mn platforms exclusively.

While at it, remove a patch that is no longer necessary, since the latest
versions of qtbase already have the patch's change in them.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>