imx-gpu-viv-v6: remove kernel-module-imx-gpu-viv dependency
This dependency was originally removed in commit
eafa8684a3ccdbad39a3fcf8cf0b6d6d79d11c5c, but it was erroneously put back in
commit 793e901678bb68f289a951be1d2e1a0f7d8face9 when updating the recipe using
the one in meta-fsl-bsp-release as reference.

Instead of removing the line containing the dependency, in order to avoid the
same situation from happening in the future,  comment it out and add an
explanation about why we're removing it.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>