imx-atf: add patch to fix the ccimx8mn hanging on suspend/resume
The LPDDR4 retention code is executed as part of the suspend/resume process.
When resuming, this code writes to a register that doesn't exist on the
i.MX8MN, causing the system to freeze before control is returned to Linux.

Add a patch that introduces a compile-time check for the SoC type to avoid
performing the write operation on the i.MX8MN.

https://jira.digi.com/browse/DEL-6914

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>