dey-image-installer: fix BOOTABLE_ARTIFACTS for ConnectCore 8M Nano platform
The bootable artifacts that must go inside the installer ZIP image, are defined
in the variable BOOTABLE_ARTIFACTS. For platforms without RAM_CONFIGS, these
artifacts are obtained from the UBOOT_CONFIG variable.

This commit fixes the final artifact name added to the BOOTABLE_ARTIFACTS that
for some platform is not strictly the same that the UBOOT_CONFIG name.
For example for the U-Boot configuration "ccimx8mn_dvk", corresponds the
bootable artifact "ccimx8mn-dvk"

https://jira.digi.com/browse/DEL-6974

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>