sdk: remove image name from default SDK path
Some configuration files inside of the SDK make use of the default SDK
installation path, so some tools might break unless the SDK is installed in said
default path.

Recently, we modified the default installation path to include the platform and
the image type, but the image type was added after the SDK was created, so even
though the environment script's paths include the image type, the "original"
default path in the config files inside of the SDK doesn't include the image
type. A side effect of this is that Qt5 apps cannot be built, since the qmake
and Qt configuration files are pointing to the "original" SDK path.

Remove the image type from the path so that the paths in the SDK's config files
match the real default installation path.

This partially reverts commit be0fe088e39a97c7dbbba62851555b17d970df79.

https://jira.digi.com/browse/DEL-6972

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>