trustfence-sign-kernel.sh: don't generate SRK_efuses for DEY sign script
The only script that needs to generate the SRK_efuses is the sign.sh
script in the U-Boot code. For the rest of signed non-bootable artifacts
this is not required and it was creating the SRK_efuses file on every
recipe where the script was called, like linux-dey and others, which
eventually resulted in a conflict when copying the artifacts to the shared
deploy-image-dir.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>