trustfence-sign-tools: homogenize the sign script filename
The recipe needs to create a copy of the sign.sh script to be used by
other recipes, but the file is the same whether you use it for HAB or AHAB
images. This is determined through the use of an exported variable with
the mode. There is no need to have the script duplicated.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>