firmware-qualcomm: fix Bluetooth NVM file customization
We are currently applying customizations to the base NVM file on install.
There was an error that introduced the '"\' characters into the NVM file
instead of the intended hex data.

This commit fixes the implementation of this customization.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>
(cherry picked from commit a204a7bfc54fc15aca0d87530c838fb427710a8e)