firmware-qualcomm: add bluetooth customizations for CC6Plus
Some CC6Plus units fail to communicate with the Bluetooth chip at a certain
temperature range, as happened on some CC6UL units.

Apply the following customizations to the NVM file so the communication is
maintained regardless of the temperature.

- Disable DEEP_SLEEP
- Enable Internal Clock

These customizations were already performed on the NVM file for the CC6UL.

https://jira.digi.com/browse/DEL-3711

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>