ccimx8x: add support for QXP overlay
Each platform now has only one base device tree, meant for the i.MX8DX variants
of the ccimx8x boards. Apply the QXP overlay in the bootscript if the SOC is
identified as an i.MX8QXP.

Since we want to give this overlay a higher priority than the SOM overlays
(wifi and bluetooth), prepend it to the overlay list after appending the SOM
overlays.

https://jira.digi.com/browse/DEL-7100

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>