ccimx6qpsbc: bootscript: hardcode imx6qp soc_family for all SOC variants
In the ccimx6sbc, since each SOC has a separate set of device trees, we can use
the soc_family environment variable when generating the name of the final
device tree in the bootscript. However, for the ccimx6qpsbc, there is only one
set of device trees with the "imx6qp" prefix, even though there are SOC
variants such as the imx6dp. Running the bootscript on a ccimx6qpsbc with an
imx6dp SOC assembled will result in a failure, since it will look for a
non-existing device tree starting with a "imx6dp" prefix.

Stop using the soc_family variable and hardcode "imx6qp" instead.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>