dey-image-recovery-initramfs: remove openssl-bin from initramfs
This package was being included as a RRECOMMENDS from libcrypto, but we don't
need it in the initramfs at all. Since our recovery image is already pretty big
compared with the recovery partition size on ccimx6ul platforms with 256 MB of
storage, remove it to make sure the image can fit.

https://jira.digi.com/browse/DEL-7253

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>