cog: fix crash due to EGL library errors
The EGL library provided by NXP in the v5.4.47 BSP has a bug where a NULL
pointer is returned by eglGetProcAddress() when trying to obtain the entrypoint
for a specific function used throughout the cog code. This caused cog to crash
immediately after launching it.

Recently, a fix was made available in the cog repo, so backport it to the
latest available revision of cog in meta-webkit and apply it.

https://jira.digi.com/browse/DEL-7311
https://jira.digi.com/browse/DEL-7312
https://jira.digi.com/browse/DEL-7339

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>