update-fw scripts: fix error path after sucessful update
On commit 1e5cafeb7945401794765ef55ccf89f95ba3361c a fall-back
update command was introduced in case the newest command fails.
If the newest command succeeds, however, the 'test' command for
error doesn't match (there is no error) and the return code of
such test is "1". On the next check for errors, this results on
an error, when we really mean to continue because the first
command was successful.

Move the second error check inside the error path of the first
attempt.

Reported-by: Francisco Gil <francisco.gilmartinez@digi.com>
Signed-off-by: Hector Palacios <hector.palacios@digi.com>