weston-init: fix build by removing duplicate entry for ccimx6 platforms
In the free meta-scale layer, the same entry "use-g2d = 1" is uncommented for
mx6dl and mx6q, however our ccimx6 platform has both machines in its
definition, so the same entry is uncommented twice and throws an unexpected
error. This commit adds a append in the install process to uncomment this entry
for the ccimx6 platform.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>